backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
groq==0.13.0
PyGithub==2.5.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
python-multipart==0.0.20
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4